+++ docs/demo/demo-editable.js
@@ -16,7 +16,7 @@ new mar10.Wunderbaum({
   id: "demo",
   element: document.getElementById("demo-tree"),
   debugLevel: 5,
-  connectTopBreadcrumb: document.getElementById("parentPath"),
+  connectTopBreadcrumb: "output#parentPath",
   checkbox: true,
   // fixedCol: true,
   // navigationModeOption: "row",
@@ -107,8 +107,13 @@ new mar10.Wunderbaum({
     },
   },
   filter: {
-    connectInput: "input#filterQuery",
-    // noData: "No matching nodes",
+    connect: {
+      inputElem: "#filter-query",
+      modeButton: "#filter-hide",
+      nextButton: "#filter-next",
+      prevButton: "#filter-prev",
+      matchInfoElem: "#filter-match-info",
+    },
   },
   init: (e) => {},
   // load: function (e) {