+++ docs/demo/demo-readonly.js
@@ -16,7 +16,7 @@ new mar10.Wunderbaum({
   id: "demo",
   element: document.getElementById("demo-tree"),
   debugLevel: 5,
-  connectTopBreadcrumb: document.getElementById("parentPath"),
+  connectTopBreadcrumb: "output#parentPath",
   // checkbox: true,
   // fixedCol: true,
   navigationModeOption: "cell",
@@ -49,7 +49,13 @@ new mar10.Wunderbaum({
   columnsResizable: true,
   columnsSortable: true,
   filter: {
-    connectInput: "input#filterQuery",
+    connect: {
+      inputElem: "#filter-query",
+      modeButton: "#filter-hide",
+      nextButton: "#filter-next",
+      prevButton: "#filter-prev",
+      matchInfoElem: "#filter-match-info",
+    },
   },
   init: (e) => {},
   buttonClick: function (e) {